requirements.txt
fastapi==0.122.0
matplotlib==3.10.7
numpy==2.3.5
Pillow==12.0.0
scikit_learn==1.7.2
torch==2.7.1
torchvision==0.22.1
